+++ src/vector3.c
@@ -14,19 +14,19 @@ void    vector3_minus(t_vector3 *a, const t_vector3 *b, const t_vector3 *c)
     a->z = b->z - c->z;
 }
 
-float   vector3_dot(const t_vector3 *a, const t_vector3 *b)
+double   vector3_dot(const t_vector3 *a, const t_vector3 *b)
 {
     return (a->x * b->x + a->y * b->y + a->z * b->z);
 }
 
-float   vector3_length(const t_vector3 *a)
+double   vector3_length(const t_vector3 *a)
 {
     return (sqrt(vector3_dot(a, a)));
 }
 
-void    vector3_div(t_vector3 *dst, const t_vector3 *src, const float len)
+void    vector3_div(t_vector3 *dst, const t_vector3 *src, const double len)
 {
-    const float inv_len = 1 / len;
+    const double inv_len = 1 / len;
 
     dst->x = src->x * inv_len;
     dst->y = src->y * inv_len;
@@ -38,7 +38,7 @@ void    vector3_normalized(t_vector3 *src)
     vector3_div(src, src, vector3_length(src));
 }
 
-void    vector3_mul(t_vector3 *dst, const t_vector3 *src, const float value)
+void    vector3_mul(t_vector3 *dst, const t_vector3 *src, const double value)
 {
     dst->x = src->x * value;
     dst->y = src->y * value;
